api/requirements-dev.txt
pytest-django==3.5.1
ipdb>=0.12.2
mock>=3.0.5